Delete 3 legacy files now dead after engine removal (-1461 lines)
Deleted:
- legacy_hsync_infra_beat.go (109 lines) — StartInfraBeatLoop, only
  called from removed StartAgent MP block
- legacy_agent_policy.go (503 lines) — EvaluateUpdate/ProcessUpdate,
  only called from SynchedDataEngine
- legacy_syncheddataengine.go (849 lines) — SynchedDataEngine and
  constructors, only called from removed StartAgent MP block

Relocated to mpmethods.go:
- NewAgentRepo(), NewZoneDataRepo() constructors
- allRecipientsConfirmed() helper

Remaining legacy files are blocked by legacy API handlers
(apihandler_agent.go, apihandler_combiner.go, etc.) which
directly call their functions.

Build verified: all tdns + tdns-mp binaries compile.
